pyproject.toml
[tool.poetry]
name = "video-poc"
version = "0.1.0"
description = "YouTube Data API video fetcher for short videos"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "services"}]

[tool.poetry.dependencies]
python = "^3.8.1"
google-api-python-client = "^2.100.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
youtube-fetch = "services.01_url_pull.main:main"